// frontend/src/api/api.ts
export interface BuiltinToolCallPart {
    /**
     * 
     * @type {string}
     * @memberof BuiltinToolCallPart
     */
    'tool_name': string;
    /**
     * 
     * @type {Args}
     * @memberof BuiltinToolCallPart
     */
    'args'?: Args | null;
    /**
     * 
     * @type {string}
     * @memberof BuiltinToolCallPart
     */
    'tool_call_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof BuiltinToolCallPart
     */
    'provider_name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BuiltinToolCallPart
     */
    'part_kind'?: BuiltinToolCallPartPartKindEnum;
}

export const BuiltinToolCallPartPartKindEnum = {
    BuiltinToolCall: 'builtin-tool-call'
} as const;

export type BuiltinToolCallPartPartKindEnum = typeof BuiltinToolCallPartPartKindEnum[keyof typeof BuiltinToolCallPartPartKindEnum];

/**
 * 
 * @export
 * @interface BuiltinToolReturnPart
 */
export interface BuiltinToolReturnPart {
    /**
     * 
     * @type {string}
     * @memberof BuiltinToolReturnPart
     */
    'tool_name': string;
    /**
     * 
     * @type {any}
     * @memberof BuiltinToolReturnPart
     */
    'content': any;
    /**
     * 
     * @type {string}
     * @memberof BuiltinToolReturnPart
     */
    'tool_call_id'?: string;
    /**
     * 
     * @type {any}
     * @memberof BuiltinToolReturnPart
     */
    'metadata'?: any;
    /**
     * 
     * @type {string}
     * @memberof BuiltinToolReturnPart
     */
    'timestamp'?: string;
    /**
     * 
     * @type {string}
     * @memberof BuiltinToolReturnPart
     */
    'provider_name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BuiltinToolReturnPart
     */
    'part_kind'?: BuiltinToolReturnPartPartKindEnum;
}

export const BuiltinToolReturnPartPartKindEnum = {
    BuiltinToolReturn: 'builtin-tool-return'
} as const;

export type BuiltinToolReturnPartPartKindEnum = typeof BuiltinToolReturnPartPartKindEnum[keyof typeof BuiltinToolReturnPartPartKindEnum];

/**
 * 
 * @export
 * @interface ChatRequest
 */
export interface ChatRequest {
    /**
     * 
     * @type {Array<ChatRequestMessagesInner>}
     * @memberof ChatRequest
     */
    'messages': Array<ChatRequestMessagesInner>;
}
/**
 * @type ChatRequestMessagesInner
 * @export
 */
export type ChatRequestMessagesInner = { kind: 'request' } & ModelRequest | { kind: 'response' } & ModelResponse;
// ...
export interface FilePart {
    /**
     * 
     * @type {BinaryContent}
     * @memberof FilePart
     */
    'content': BinaryContent;
    /**
     * 
     * @type {string}
     * @memberof FilePart
     */
    'id'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FilePart
     */
    'provider_name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FilePart
     */
    'part_kind'?: FilePartPartKindEnum;
}
// ...
export interface ModelRequest {
    /**
     * 
     * @type {Array<ModelRequestPartsInner>}
     * @memberof ModelRequest
     */
    'parts': Array<ModelRequestPartsInner>;
    /**
     * 
     * @type {string}
     * @memberof ModelRequest
     */
    'instructions'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ModelRequest
     */
    'kind'?: ModelRequestKindEnum;
}

export const ModelRequestKindEnum = {
    Request: 'request'
} as const;

export type ModelRequestKindEnum = typeof ModelRequestKindEnum[keyof typeof ModelRequestKindEnum];

/**
 * @type ModelRequestPartsInner
 * @export
 */
export type ModelRequestPartsInner = { part_kind: 'retry-prompt' } & RetryPromptPart | { part_kind: 'system-prompt' } & SystemPromptPart | { part_kind: 'tool-return' } & ToolReturnPart | { part_kind: 'user-prompt' } & UserPromptPart;

/**
 * 
 * @export
 * @interface ModelResponse
 */
export interface ModelResponse {
    /**
     * 
     * @type {Array<ModelResponsePartsInner>}
     * @memberof ModelResponse
     */
    'parts': Array<ModelResponsePartsInner>;
    /**
     * 
     * @type {RequestUsage}
     * @memberof ModelResponse
     */
    'usage'?: RequestUsage;
    /**
     * 
     * @type {string}
     * @memberof ModelResponse
     */
    'model_name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ModelResponse
     */
    'timestamp'?: string;
    /**
     * 
     * @type {string}
     * @memberof ModelResponse
     */
    'kind'?: ModelResponseKindEnum;
    /**
     * 
     * @type {string}
     * @memberof ModelResponse
     */
    'provider_name'?: string | null;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof ModelResponse
     */
    'provider_details'?: { [key: string]: any; } | null;
    /**
     * 
     * @type {string}
     * @memberof ModelResponse
     */
    'provider_response_id'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ModelResponse
     */
    'finish_reason'?: ModelResponseFinishReasonEnum | null;
}

export const ModelResponseKindEnum = {
    Response: 'response'
} as const;

export type ModelResponseKindEnum = typeof ModelResponseKindEnum[keyof typeof ModelResponseKindEnum];
export const ModelResponseFinishReasonEnum = {
    Stop: 'stop',
    Length: 'length',
    ContentFilter: 'content_filter',
    ToolCall: 'tool_call',
    Error: 'error'
} as const;

export type ModelResponseFinishReasonEnum = typeof ModelResponseFinishReasonEnum[keyof typeof ModelResponseFinishReasonEnum];

/**
 * @type ModelResponsePartsInner
 * @export
 */
export type ModelResponsePartsInner = { part_kind: 'builtin-tool-call' } & BuiltinToolCallPart | { part_kind: 'builtin-tool-return' } & BuiltinToolReturnPart | { part_kind: 'file' } & FilePart | { part_kind: 'text' } & TextPart | { part_kind: 'thinking' } & ThinkingPart | { part_kind: 'tool-call' } & ToolCallPart;

/**
 * 
 * @export
 * @interface RequestUsage
 */
export interface RequestUsage {
    /**
     * 
     * @type {number}
     * @memberof RequestUsage
     */
    'input_tokens'?: number;
    /**
     * 
     * @type {number}
     * @memberof RequestUsage
     */
    'cache_write_tokens'?: number;
    /**
     * 
     * @type {number}
     * @memberof RequestUsage
     */
    'cache_read_tokens'?: number;
    /**
     * 
     * @type {number}
     * @memberof RequestUsage
     */
    'output_tokens'?: number;
    /**
     * 
     * @type {number}
     * @memberof RequestUsage
     */
    'input_audio_tokens'?: number;
    /**
     * 
     * @type {number}
     * @memberof RequestUsage
     */
    'cache_audio_read_tokens'?: number;
    /**
     * 
     * @type {number}
     * @memberof RequestUsage
     */
    'output_audio_tokens'?: number;
    /**
     * 
     * @type {{ [key: string]: number; }}
     * @memberof RequestUsage
     */
    'details'?: { [key: string]: number; };
}
/**
 * 
 * @export
 * @interface RetryPromptPart
 */
export interface RetryPromptPart {
    /**
     * 
     * @type {Content}
     * @memberof RetryPromptPart
     */
    'content': Content;
    /**
     * 
     * @type {string}
     * @memberof RetryPromptPart
     */
    'tool_name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RetryPromptPart
     */
    'tool_call_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof RetryPromptPart
     */
    'timestamp'?: string;
    /**
     * 
     * @type {string}
     * @memberof RetryPromptPart
     */
    'part_kind'?: RetryPromptPartPartKindEnum;
}

export const RetryPromptPartPartKindEnum = {
    RetryPrompt: 'retry-prompt'
} as const;

export type RetryPromptPartPartKindEnum = typeof RetryPromptPartPartKindEnum[keyof typeof RetryPromptPartPartKindEnum];

/**
 * 
 * @export
 * @interface SystemPromptPart
 */
export interface SystemPromptPart {
    /**
     * 
     * @type {string}
     * @memberof SystemPromptPart
     */
    'content': string;
    /**
     * 
     * @type {string}
     * @memberof SystemPromptPart
     */
    'timestamp'?: string;
    /**
     * 
     * @type {string}
     * @memberof SystemPromptPart
     */
    'dynamic_ref'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SystemPromptPart
     */
    'part_kind'?: SystemPromptPartPartKindEnum;
}

export const SystemPromptPartPartKindEnum = {
    SystemPrompt: 'system-prompt'
} as const;

export type SystemPromptPartPartKindEnum = typeof SystemPromptPartPartKindEnum[keyof typeof SystemPromptPartPartKindEnum];

/**
 * 
 * @export
 * @interface TextPart
 */
export interface TextPart {
    /**
     * 
     * @type {string}
     * @memberof TextPart
     */
    'content': string;
    /**
     * 
     * @type {string}
     * @memberof TextPart
     */
    'id'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TextPart
     */
    'part_kind'?: TextPartPartKindEnum;
}

export const TextPartPartKindEnum = {
    Text: 'text'
} as const;

export type TextPartPartKindEnum = typeof TextPartPartKindEnum[keyof typeof TextPartPartKindEnum];

/**
 * 
 * @export
 * @interface ThinkingPart
 */
export interface ThinkingPart {
    /**
     * 
     * @type {string}
     * @memberof ThinkingPart
     */
    'content': string;
    /**
     * 
     * @type {string}
     * @memberof ThinkingPart
     */
    'id'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ThinkingPart
     */
    'signature'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ThinkingPart
     */
    'provider_name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ThinkingPart
     */
    'part_kind'?: ThinkingPartPartKindEnum;
}

export const ThinkingPartPartKindEnum = {
    Thinking: 'thinking'
} as const;

export type ThinkingPartPartKindEnum = typeof ThinkingPartPartKindEnum[keyof typeof ThinkingPartPartKindEnum];

/**
 * 
 * @export
 * @interface ToolCallPart
 */
export interface ToolCallPart {
    /**
     * 
     * @type {string}
     * @memberof ToolCallPart
     */
    'tool_name': string;
    /**
     * 
     * @type {Args}
     * @memberof ToolCallPart
     */
    'args'?: Args | null;
    /**
     * 
     * @type {string}
     * @memberof ToolCallPart
     */
    'tool_call_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ToolCallPart
     */
    'part_kind'?: ToolCallPartPartKindEnum;
}

export const ToolCallPartPartKindEnum = {
    ToolCall: 'tool-call'
} as const;

export type ToolCallPartPartKindEnum = typeof ToolCallPartPartKindEnum[keyof typeof ToolCallPartPartKindEnum];

/**
 * 
 * @export
 * @interface ToolReturnPart
 */
export interface ToolReturnPart {
    /**
     * 
     * @type {string}
     * @memberof ToolReturnPart
     */
    'tool_name': string;
    /**
     * 
     * @type {any}
     * @memberof ToolReturnPart
     */
    'content': any;
    /**
     * 
     * @type {string}
     * @memberof ToolReturnPart
     */
    'tool_call_id'?: string;
    /**
     * 
     * @type {any}
     * @memberof ToolReturnPart
     */
    'metadata'?: any;
    /**
     * 
     * @type {string}
     * @memberof ToolReturnPart
     */
    'timestamp'?: string;
    /**
     * 
     * @type {string}
     * @memberof ToolReturnPart
     */
    'part_kind'?: ToolReturnPartPartKindEnum;
}

export const ToolReturnPartPartKindEnum = {
    ToolReturn: 'tool-return'
} as const;

export type ToolReturnPartPartKindEnum = typeof ToolReturnPartPartKindEnum[keyof typeof ToolReturnPartPartKindEnum];

/**
 * 
 * @export
 * @interface UserPromptPart
 */
export interface UserPromptPart {
    /**
     * 
     * @type {Content1}
     * @memberof UserPromptPart
     */
    'content': Content1;
    /**
     * 
     * @type {string}
     * @memberof UserPromptPart
     */
    'timestamp'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserPromptPart
     */
    'part_kind'?: UserPromptPartPartKindEnum;
}